sign-inspire-backend/requirements.txt
fastapi[standard]
uvicorn
apscheduler
langchain
langchain-openai
langchain-google-genai
pydantic
sqlalchemy
pymysql
cryptography
python-dotenv
httpx
pypinyin
